requirements.txt
beautifulsoup4==4.6.0
bleach==1.5.0
boilerpipe3==1.1
boto==2.48.0
bs4==0.0.1
bz2file==0.98
certifi==2017.11.5
charade==1.0.3
chardet==3.0.4
click==6.7
enum34==1.1.6
gensim==3.1.0
graphviz==0.8.1
h5py==2.7.1
html5lib==0.9999999
idna==2.6
JPype1-py3==0.5.5.2
Keras==2.0.9
Markdown==2.6.9
nltk==3.2.5
numpy==1.13.3
protobuf==3.4.0
pydot==1.2.3
pyparsing==2.2.0
python-dotenv==0.7.1
PyYAML==3.12
requests==2.18.4
scikit-learn==0.19.1
scipy==1.0.0
six==1.11.0
sklearn==0.0
smart-open==1.5.3
tensorflow-tensorboard==0.4.0rc2
tqdm==4.19.4
urllib3==1.22
Werkzeug==0.12.2
